chore(GEM-43): add back settings page
